IMP-1158: fix [CHANGE REQUEST] We need to be able to top up users' accounts in JAdmin to support the new monthly billing model
Conflicts:
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/User.java
	service/src/main/java/mobi/nowtechnologies/server/assembler/UserAsm.java
	shared/src/main/java/mobi/nowtechnologies/server/shared/Utils.java

+++ persistence/src/test/java/mobi/nowtechnologies/server/persistence/domain/UserTest.java
@@ -44,7 +44,7 @@ public class UserTest {
 	@Test
 	public void isOnFreeTrial_false_when_freeTrialExpiredMillis_Eq_Null() {
 		User user = new User();
-			user.setFreeTrialExpiredMillis(null);
+			user.setFreeTrialExpiredMillis((Long)null);
 		assertEquals(false, user.isOnFreeTrial());
 	}
